Add parser benchmark.
